feat: implement action system for bot actions with LoginAction and VoteAction, and integrate voting functionality
