+++ plugin/src/main/java/me/seetaadev/serverfiller/bot/service/actions/Action.java
@@ -1,0 +1,17 @@
+package me.seetaadev.serverfiller.bot.service.actions;
+
+import java.util.Random;
+
+public interface Action {
+
+    Random rand = new Random();
+    void start();
+    void stop();
+
+    default int getRandomDelayTicks(int min, int max) {
+        int range = Math.max(max - min + 1, 1);
+        return (min + rand.nextInt(range)) * 20;
+    }
+
+    default void ensureMinimum() {}
+}
