+++ plugin/src/main/java/me/seetaadev/serverfiller/hooks/HookManager.java
@@ -1,14 +1,22 @@
 package me.seetaadev.serverfiller.hooks;
 
+import me.seetaadev.serverfiller.ServerFillerPlugin;
 import me.seetaadev.serverfiller.bot.Bot;
 import me.seetaadev.serverfiller.hooks.discordsrv.DiscordSRVHook;
 import me.seetaadev.serverfiller.hooks.luckperms.LuckPermsHook;
+import me.seetaadev.serverfiller.hooks.voting.VotingHook;
 import org.bukkit.Bukkit;
 
 public class HookManager {
 
+    private final ServerFillerPlugin plugin;
     private DiscordSRVHook discordSRVHook;
     private LuckPermsHook luckPermsHook;
+    private VotingHook votingHook;
+
+    public HookManager(ServerFillerPlugin plugin) {
+        this.plugin = plugin;
+    }
 
     public void init() {
         if (Bukkit.getPluginManager().isPluginEnabled("DiscordSRV")) {
@@ -17,6 +25,10 @@ public class HookManager {
 
         if (Bukkit.getPluginManager().isPluginEnabled("LuckPerms")) {
             luckPermsHook = new LuckPermsHook();
+        }
+
+        if (Bukkit.getPluginManager().isPluginEnabled("VotingPlugin")) {
+            votingHook = new VotingHook(plugin);
         }
     }
 
@@ -30,5 +42,9 @@ public class HookManager {
         if (luckPermsHook != null) {
             luckPermsHook.giveRank(name, rank);
         }
+    }
+
+    public void vote(Bot bot) {
+        votingHook.vote(bot);
     }
 }
